Fix NULL use, some const params, and don't use implicit bools.

+++ Reactions/ReactorCvodeJacobian.cpp
@@ -130,7 +130,8 @@ cJac(
 
     // Jac
     amrex::Real Jmat_tmp[(NUM_SPECIES + 1) * (NUM_SPECIES + 1)] = {0.0};
-    const int consP = reactor_type == ReactorTypes::h_reactor_type;
+    const int consP =
+      static_cast<int>(reactor_type == ReactorTypes::h_reactor_type);
     auto eos = pele::physics::PhysicsType::eos();
     eos.RTY2JAC(rho, temp, massfrac, Jmat_tmp, consP);
 
@@ -217,7 +218,8 @@ cJac_sps(
     // Do we recompute Jac ?
     amrex::Real Jmat_tmp[(NUM_SPECIES + 1) * (NUM_SPECIES + 1)] = {0.0};
     if (fabs(temp - temp_save_lcl) > 1.0) {
-      const int consP = reactor_type == ReactorTypes::h_reactor_type;
+      const int consP =
+        static_cast<int>(reactor_type == ReactorTypes::h_reactor_type);
       auto eos = pele::physics::PhysicsType::eos();
       eos.RTY2JAC(rho, temp, massfrac, Jmat_tmp, consP);
       temp_save_lcl = temp;
